Move component files
