+++ src/app/modules/game/components/card-list.component.css
