+++ src/app/modules/game/components/card-list.component.html
