+++ src/app/modules/game/components/card.component.css
