+++ src/app/modules/game/components/flip-result.component.spec.ts
