+++ src/app/modules/game/components/game-progress.component.html
