+++ src/app/modules/game/components/game.component.ts
@@ -9,10 +9,10 @@ import {
   effect,
 } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { FlipResultComponent } from '../../components/flip-result/flip-result.component';
-import { GAME_DIFFICULTY } from '../../constants/game-difficulty';
-import { GameService } from '../../services/game.service';
-import type { Card, GameStatus } from '../../types';
+import { FlipResultComponent } from './flip-result.component';
+import { GAME_DIFFICULTY } from '../constants/game-difficulty';
+import { GameService } from '../services/game.service';
+import type { Card, GameStatus } from '../types';
 
 @Component({
   selector: 'app-game',
