+++ src/app/modules/game/components/top-title.component.ts
