+++ src/app/modules/game/components/top.component.html
